test-requirements.txt
boto3
click
docker
flask
requests
schedule
titus-optimize==0.7.2
cvxopt==1.2.5
netflix-spectator-py
kubernetes
grpcio
grpcio-tools
